requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
albumentations>=1.0.3
labelme>=4.5.9
scikit-image>=0.18.3
tqdm>=4.62.3
opencv-python>=4.5.3
